requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
celery
redis
python-decouple
gunicorn
uvicorn
drf-spectacular
